csv_chat/requirements.txt
langchain
openai
python-dotenv
streamlit
tabulate